refactor(ui): extract form controls to shared EditorUiComponents library
Move PushControlStyle, SettingRow, SettingGroup, ComboControl,
SliderIntControl, ToggleControl, InputTextControl, InputIntControl,
InputFloatControl, PluginRow, ShortcutDisplay, ThemeChip from
SettingsModal.cpp anonymous namespace into Horo::Editor::Ui.

New header: include/Horo/Editor/EditorUiComponents.h
Implementation: src/editor/design_system/components/EditorUiComponents.cpp

SettingsModal now imports via using namespace Ui.
NewProjectModal and WelcomeScreen already used Ui:: primitives.
Added Theme::Layout::ControlW for shared form column width.
